move the drop-acl to acl.go and add insert line number
